+++ gnssrefl/quickplt.py
@@ -11,23 +11,60 @@ import sys
 import gnssrefl.quicklib as q
 import gnssrefl.gps as g
 
+from gnssrefl.utils import validate_input_datatypes, str2bool
 
-def main():
+def parse_arguments():
+    parser = argparse.ArgumentParser()
+
+    parser.add_argument("filename", help="filename", type=str)
+    parser.add_argument("xcol", help="x-column", type=int)
+    parser.add_argument("ycol",   help="y-column", type=int)
+    parser.add_argument("-errorcol",   help="error bar for y-values", type=int,default=None)
+    parser.add_argument("-mjd", help="if x-values are MJD ", type=str,default=None)
+    parser.add_argument("-reverse", help="reverse the y-axis", type=str,default=None)
+    parser.add_argument("-ymdhm", help="columns 1-5 are year mon day hour minute", type=str,default=None)
+    parser.add_argument("-xlabel", type=str, help="optional x-axis label", default=None)
+    parser.add_argument("-ylabel", type=str, help="optional y-axis label", default=None)
+    parser.add_argument("-symbol", help="plot symbol ", type=str,default=None)
+    parser.add_argument("-title", help="optional title", type=str,default=None)
+    parser.add_argument("-outfile", help="optional filename for plot. Must end in png", type=str,default=None)
+    parser.add_argument("-xlimits", nargs="*",type=float, help="optional xlimits", default=None)
+    parser.add_argument("-ylimits", nargs="*",type=float, help="optional ylimits", default=None)
+    parser.add_argument("-ydoy", help="if True/T, columns 1-2 are year and doy", type=str,default=None)
+    parser.add_argument("-filename2", help="second filename", type=str, default=None)
+    parser.add_argument("-freq", help="spec freq, column 11 ", type=int,default=None)
+    parser.add_argument("-utc_offset", help="offset from UTC, hours  ", type=int,default=None)
+
+
+    args = parser.parse_args().__dict__
+
+    # convert all expected boolean inputs from strings to booleans
+    boolean_args = ['mjd', 'reverse', 'ymdhm', 'ydoy']
+    args = str2bool(args, boolean_args)
+
+    # only return a dictionary of arguments that were added from the user - all other defaults will be set in code below
+    return {key: value for key, value in args.items() if value is not None}
+
+
+def run_quickplt (filename: str, xcol: int, ycol: int, errorcol: int=None, mjd: bool=False, xlabel: str=None, 
+                  ylabel: str=None, symbol: str=None, reverse:bool=False,title:str=None,outfile: str=None,
+                  xlimits: float=[], ylimits: float=[], ydoy:bool=False, ymdhm:bool=False, filename2: str=None, freq:int=None, utc_offset: int=None):
+
     """
     quick file plotting using matplotlib
 
     A png file is saved as temp.png or to your preferred
     filename if outfile is given. In either case, it goes to REFL_CODE/Files
 
-    Allows you to set x and y-axis limits with a title and various axes labels
-
-    a different symbol can also be set
+    Allows you to set x and y-axis limits with a title and various axes labels, symbols etc.
     
     Someone could easily update this to include different filetypes (e.g. jpeg)
 
+    I rewrote this recently to take advantage of our boolean argument translator.  Let me know 
+    if things have broken or submit a PR.
+
     Examples
     --------
-
     quickplt txtfile 1 16
         would plot column 1 on the x-axis and column 16 on the y-axis
 
@@ -45,6 +82,16 @@ def main():
         it would reverse the y-axis parameter as you might want if 
         you are ploting RH but want it to have the same sense as a tide gauge.
 
+    quickplt txtfile 1 3 -errorcol 4 
+        would plot error bars from column 4 
+
+    quickplt txtfile 1 3 -errorcol 4 -ydoy T
+        would plot error bars from column 4  and assume columns 1 and 2 
+        are year and day of year
+
+    quickplt txtfile 1 4 -mjd T
+        assume column 1 is modified julian day
+
     quickplt txtfile 1 16 -ylimits 0 2
         would restrict y-axis to be between 0 and 2
 
@@ -56,97 +103,66 @@ def main():
     ----------
     filename : str
         name of file to be plotted 
-    xcol : str
+    xcol : int
         column number in the file for the x-axis parameter
-    ycol : str
+    ycol : int
         column number in the file for the y-axis parameter
-    mjd : str
-        T or True, code will convert MJD to datetime, optional
+    errorcol : int, optional
+        column number for the error bars
+    mjd : bool, optional
+        code will convert MJD to datetime (for xcol) 
     xlabel : str, optional
         label for x-axis 
     ylabel : str
         label for y-axis 
     symbol : str, optional
-        prescibe the marker used in the plot 
-    reverse : str, optional
-        T or True, to reverse y-axis limits
+        prescribe the marker used in the plot . It can include the color, i.e.
+        'b.' or 'b^'
+    reverse : bool, optional
+        to reverse y-axis limits
     title : str, optional
         title for plot 
     outfile : str, optional
         name of png file to store plot 
-    xlimits: float, optional
-        pair of xaxis limits  
-    ylimits: float, optional
-        pair of yaxis limits  
+    xlimits: list of floats, optional
+        xaxis limits  
+    ylimits: list of floats, optional
+        yaxis limits  
     ydoy : bool, optional
         if columns 1 and 2 are year and doy, the x-axis will be plotted in obstimes
         you should select column 1 to plot
+    ymdhm : bool, optional
+        if columns 1-5 are Y,M,D,H,M then x-axis will be plotted in obstimes
     filename2 : str
         in principle this allows you to make plots from two files with identical formatting
         not sure that it works
-    freq: integer, optional
+    freq: int, optional
         use column 11 to find (and extract) a single frequency
-    utc_offset: integer, optional
+    utc_offset: int, optional
         offset time axis by this number of hours (for local time)
         this only is used when the mjd option is used  
 
     """
 
-    parser = argparse.ArgumentParser()
-    parser.add_argument("filename", help="filename", type=str)
-    parser.add_argument("xcol", help="x-column", type=str)
-    parser.add_argument("ycol",   help="y-column", type=str)
-    parser.add_argument("-mjd", help="set to True/T if x-values are MJD ", type=str,default=None)
-    parser.add_argument("-reverse", help="set to True/T to reverse the y-axis", type=str,default=None)
-    parser.add_argument("-ymdhm", help="if True/T, columns 1-5 are year mon day hour minute", type=str,default=None)
-    parser.add_argument("-xlabel", type=str, help="optional x-axis label", default=None)
-    parser.add_argument("-ylabel", type=str, help="optional y-axis label", default=None)
-    parser.add_argument("-symbol", help="plot symbol ", type=str,default=None)
-    parser.add_argument("-title", help="optional title", type=str,default=None)
-    parser.add_argument("-outfile", help="optional filename for plot. Must end in png", type=str,default=None)
-    parser.add_argument("-xlimits", nargs="*",type=float, help="optional xlimits", default=None)
-    parser.add_argument("-ylimits", nargs="*",type=float, help="optional ylimits", default=None)
-    parser.add_argument("-ydoy", help="if True/T, columns 1-2 are year and doy", type=str,default=None)
-    parser.add_argument("-filename2", help="second filename", type=str, default=None)
-    parser.add_argument("-freq", help="spec freq, column 11 ", type=int,default=None)
-    parser.add_argument("-utc_offset", help="offset from UTC, hours  ", type=int,default=None)
-
-    args = parser.parse_args()
-
-    filename = args.filename
     # change column numbers to pythonese
-    xcol = int(args.xcol) - 1
-    ycol = int(args.ycol) - 1
-
-
-    if args.ylabel is None:
-        ylabel = 'Unknown'
+    xcol = xcol - 1
+    ycol = ycol - 1
+    if errorcol is None:
+        yerrors = False
     else:
-        ylabel = args.ylabel
+        yerrors = True
+        errorcol = errorcol - 1
+
+    if ylabel is None:
+        ylabel = 'Unknown'
 
     secondFile = False
 
-    reverse_sign = False
-    if (args.reverse == 'True') or (args.reverse == 'T'):
-        reverse_sign = True
+    reverse_sign = reverse
 
-    # was previously ymdh
-    ymd = False
-    if (args.ymdhm == 'True') or (args.ymdhm == 'T'):
-        ymd = True
+    ymd = ymdhm
 
-    ydoy = False
-    if (args.ydoy == 'True') or (args.ydoy == 'T'):
-        ydoy = True
-
-    if args.utc_offset is not None:
-        utc_offset = args.utc_offset
-    else:
-        utc_offset = None
-
-    convert_mjd = False
-    if (args.mjd== 'True') or (args.mjd == 'T'):
-        convert_mjd = True
+    convert_mjd = mjd
 
     commentsign = '%'
 
@@ -157,9 +173,9 @@ def main():
             return
         else:
 
-            if (args.freq is not None):
-                print('restricting frequency',args.freq)
-                ii = (tvd[:,10] == args.freq)
+            if (freq is not None):
+                print('restricting frequency',freq)
+                ii = (tvd[:,10] == freq)
                 if (len(tvd[ii,10]) == 0):
                     print('nothing for that frequency')
                     return
@@ -174,8 +190,7 @@ def main():
 
     tvd2=[]
     secondFile = False
-    if args.filename2 is not None:
-        filename2 = args.filename2
+    if filename2 is not None:
         if os.path.isfile(filename2):
             tvd2 = np.loadtxt(filename2,comments=commentsign)
             if len(tvd2) == 0:
@@ -193,43 +208,45 @@ def main():
 
     # supercedes previous trans_time ... 
     if ydoy:
-        print('Making obstimes for ydoy x-axis')
+        #print('Making obstimes for ydoy x-axis')
         tval = g.ydoy2datetime(tvd[:,0], tvd[:,1])
-
 
     fig,ax=plt.subplots()
 
-    if args.symbol is None:
-        ax.plot(tval, yval, 'b.')
+    # i.e. using default
+    if symbol is None:
+        if yerrors:
+            ax.errorbar(tval, yval, yerr=tvd[:,errorcol], fmt='.',color='blue')
+        else:
+            ax.plot(tval, yval, 'b.')
     else:
-        ax.plot(tval, yval, args.symbol)
+        if yerrors:
+            ax.errorbar(tval, yval, yerr=tvd[:,errorcol],fmt=symbol)
+        else:
+            ax.plot(tval, yval, symbol)
 
     # is second file currently supported???
     if secondFile:
         ax.plot(tval2, yval2, 'r.')
 
     plt.grid()
-
     plt.ylabel(ylabel)
 
-    if args.xlabel is not None:
-        plt.xlabel(str(args.xlabel))
+    if xlabel is not None:
+        plt.xlabel(xlabel)
 
-
-    if args.title is None:
+    if title is None:
         ax.set_title(os.path.basename(filename) )
     else:
-        ax.set_title(args.title )
+        ax.set_title(title)
 
 
 
-    if args.ylimits is not None:
+    if len(ylimits) == 2:
         print('found y-axis limits')
-        ylimits = args.ylimits
         plt.ylim((ylimits))
-    if args.xlimits is not None:
+    if len(xlimits) == 2:
         print('found x-axis limits')
-        xlimits = args.xlimits
         if convert_mjd:
             t1 = Time(xlimits[0],format='mjd')
             t1_utc = t1.utc # change to UTC
@@ -255,29 +272,13 @@ def main():
         ax.invert_yaxis()
     fig.autofmt_xdate() # obstimes
 
-    out = args.outfile
-
-    q.save_plot(out)
+    q.save_plot(outfile)
     plt.show()
-    return
-
-    xdir = os.environ['REFL_CODE']  + '/Files/' 
-    if not os.path.exists(xdir) :
-        subprocess.call(['mkdir', xdir])
-
-    if out is None:
-        out = 'temp.png'
-        print('Plotfile saved to: ', xdir + out)
-        plt.savefig(xdir + out ,dpi=300)
-    else:
-        if out[-3:] == 'png':
-            print('Plotfile saved to: ', xdir  + out)
-            plt.savefig(xdir + out,dpi=300)
-        else:
-            print('Output filename must end in png. No file is written')
 
 
-
+def main():
+    args = parse_arguments()
+    run_quickplt(**args)
 
 if __name__ == "__main__":
     main()