+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2.0.1
+
+trying to fix readthedocs formatting
+
 ## 2.0.0
 January 4, 2024